LPD-56161 Organize types on MP
