+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublisherDashboard/pages/Apps/AppCreationFlow/ProvideAppBuildPage/components/OfferingTypeCheckbox.tsx
@@ -5,6 +5,12 @@
 
 import {CheckboxCard} from '../../../../../../../components/CheckboxCard/CheckboxCard';
 
+type OfferingType = {
+	description: string;
+	disabled?: boolean;
+	label: string;
+};
+
 type OfferingTypeCheckboxProps = {
 	handleSelectCheckbox: (label: string) => void;
 	offeringTypes: OfferingType[];
