+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/types/commerce.d.ts
@@ -1,0 +1,36 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+type CommerceAccount = {
+	active: boolean;
+	logoURL: string;
+	taxId: string;
+} & Omit<Account, 'description'>;
+
+type CommerceOption = {
+	id: number;
+	key: string;
+	name: string;
+};
+
+type Catalog = {
+	accountId: number | null;
+	currencyCode: string;
+	defaultLanguageId: string;
+	externalReferenceCode: string;
+	id: number;
+	name: string;
+	system: boolean;
+};
+
+type Channel = {
+	channelId: number;
+	currencyCode: string;
+	externalReferenceCode: string;
+	id: number;
+	name: string;
+	siteGroupId: number;
+	type: string;
+};
